Fix NaN chart crash on empty data, and Decimal-as-string in analytics JSON
Two real bugs surfaced by actually using the app:

1. LineChart crashed with 'Received NaN for the y1 attribute' whenever it
   rendered before its fetch resolved — Math.max of an empty array is
   -Infinity, so every y coordinate became NaN. Charts now treat empty data as
   a normal loading state rather than an error. Also guarded the all-identical-
   values case, where max-min is 0 and the domain collapses to zero width,
   producing the same NaN a different way.

2. Analytics returned "avg_salary": "118322" as a STRING. Postgres ROUND()
   yields numeric -> psycopg Decimal -> not JSON-serializable -> stringified.
   The frontend's .toLocaleString() then silently did nothing (no thousands
   separators) and arithmetic relied on JS coercion. Cast to ::float in SQL so
   it's fixed once at the source instead of in every consumer.

Verified: overview now returns 118322.0 as a number; tsc and next build clean.

+++ services/jobs-service/app/routers/analytics.py
@@ -7,6 +7,13 @@ from fastapi import APIRouter
 
 from app.routers.jobs import _rows, cached
 
+# NOTE on the ::float casts below.
+# Postgres ROUND() on a numeric returns `numeric`, which psycopg maps to Python's Decimal.
+# Decimal isn't JSON-serializable, so it gets stringified — the API then returns
+# "avg_salary": "118322" instead of 118322, and the frontend's .toLocaleString() silently
+# does nothing to a string (no thousands separators, and any arithmetic relies on JS
+# coercion). Casting in SQL fixes it once, at the source, rather than in every consumer.
+
 router = APIRouter(prefix="/analytics", tags=["analytics"])
 
 
@@ -18,8 +25,8 @@ def overview():
         """
         SELECT COUNT(*)                                   AS total_postings,
                COUNT(DISTINCT company_id)                 AS total_companies,
-               ROUND(AVG(salary))                         AS avg_salary,
-               ROUND(AVG(CASE WHEN remote THEN 1 ELSE 0 END) * 100, 1) AS remote_percent
+               ROUND(AVG(salary))::float                  AS avg_salary,
+               ROUND(AVG(CASE WHEN remote THEN 1 ELSE 0 END) * 100, 1)::float AS remote_percent
         FROM analytics.fact_job_posting
         """
     )[0]
@@ -47,7 +54,7 @@ def salary_by_seniority():
     return _rows(
         """
         SELECT seniority,
-               ROUND(AVG(salary))   AS avg_salary,
+               ROUND(AVG(salary))::float AS avg_salary,
                COUNT(*)             AS postings
         FROM analytics.fact_job_posting
         WHERE salary IS NOT NULL
@@ -63,7 +70,7 @@ def salary_by_region():
     return _rows(
         """
         SELECT region,
-               ROUND(AVG(salary))   AS avg_salary,
+               ROUND(AVG(salary))::float AS avg_salary,
                COUNT(*)             AS postings
         FROM analytics.fact_job_posting
         WHERE salary IS NOT NULL
@@ -105,8 +112,8 @@ def skill_premium(limit: int = 15):
         )
         SELECT s.skill_name,
                COUNT(*)                                        AS postings,
-               ROUND(AVG(f.salary))                            AS avg_salary,
-               ROUND(AVG(f.salary) - (SELECT avg_all FROM overall)) AS premium_vs_average
+               ROUND(AVG(f.salary))::float                     AS avg_salary,
+               ROUND(AVG(f.salary) - (SELECT avg_all FROM overall))::float AS premium_vs_average
         FROM analytics.bridge_posting_skill b
         JOIN analytics.dim_skill s       ON s.skill_id = b.skill_id
         JOIN analytics.fact_job_posting f ON f.posting_id = b.posting_id
